Make acceptance tests executable with new TCK
